pyproject.toml
[project]
name = "tobias-voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit>=1.1.6",
    "requests>=2.33.1",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "workspace-mcp>=1.19.0",
    "openai>=2.32.0",
]
